reworking headless mode part 2
